test: Fold GraphQL parser limit coverage into ServerFunctionContainerIT
